Simplify react-select stuff
