+++ frontend/src/components/SeasonReactSelect.js
@@ -1,8 +1,6 @@
-import React, { useMemo, useContext } from 'react';
+import React, { useMemo } from 'react';
 import { toSeasonString } from '../utilities';
-import { StyledSelect } from './StyledComponents';
-import { ThemeContext } from 'styled-components';
-import { selectStyles } from '../queries/Constants';
+import CustomSelect from './CustomSelect';
 
 /**
  * Render season dropdown
@@ -28,17 +26,12 @@ function SeasonReactSelect({ cur_season, season_codes, onSeasonChange }) {
     return season_options_temp;
   }, [season_codes]);
 
-  const theme = useContext(ThemeContext);
-  const select_styles = selectStyles(theme);
-
   return (
-    <StyledSelect
-      classNamePrefix={'Select'}
+    <CustomSelect
       value={{
         value: cur_season,
         label: toSeasonString(cur_season)[0],
       }}
-      styles={select_styles}
       isSearchable={false}
       options={season_options}
       onChange={(option) => {
